Avatar-editing has no more modal.
