+++ liwords-ui/src/profile/profile.tsx
@@ -13,7 +13,6 @@ import { useLoginStateStoreContext } from '../store/store';
 import { GameMetadata, RecentGamesResponse } from '../gameroom/game_info';
 import { GamesHistoryCard } from './games_history';
 import { UsernameWithContext } from '../shared/usernameWithContext';
-import { preferredSortOrder, setPreferredSortOrder } from '../store/constants';
 import { moderateUser } from '../mod/moderate';
 
 type ProfileResponse = {
